organize actions and queries into folders
