+++ src/lib/queries/reward/getRewards.ts
@@ -1,9 +1,12 @@
 import {useQuery} from '@tanstack/react-query'
 
-import {queryKeys} from './keys'
+import {queryKeys} from '../keys'
 
 import getServerQueryClient from '@/server/utils/queryClient'
-import {getRewards, type GetRewardsParams} from '@/server/actions/getReward'
+import {
+  getRewards,
+  type GetRewardsParams,
+} from '@/server/actions/reward/getReward'
 
 const getQueryOptions = (params: GetRewardsParams) => ({
   queryKey: queryKeys.rewards.all(params).queryKey,
